Refactor codebase into modular architecture and enhance git status indicators
Major refactoring and UI improvements:
- Split monolithic 1,536-line bin/gh-here.js into 5 focused modules in lib/
- Enhanced git status display with clean dot indicators in dedicated column
- Fixed untracked file detection for files within untracked directories
- Improved visual consistency with GitHub-style folder icons and status colors
- Added proper line spacing consistency across editor components
- Streamlined table layout with minimal git status column (10px width)

Technical improvements:
- Modularized git operations, file utilities, renderers, and server routes
- Enhanced git status processing to inherit directory status for nested files
- Fixed octicons scoping issues and async/await inconsistencies
- Updated CSS for proper table column alignment and git status positioning
- Maintained backward compatibility for non-git directories
